+++ backend/migrations/20221023162721_remove_iscancel_resumejob.up.sql
@@ -1,0 +1,3 @@
+-- Add up migration script here
+ALTER TABLE resume_job DROP COLUMN is_cancel;
+ALTER TABLE resume_job ADD COLUMN approver VARCHAR(50);