+++ frontend/src/lib/components/landing/FlowGettingStarted.svelte
@@ -42,6 +42,7 @@
 			variant="border"
 			size="xs"
 			startIcon={{ icon: faExternalLink }}
+			target="_blank"
 		>
 			Explore community flows on WindmillHub
 		</Button>